requirements.txt
absl-py==0.12.0
alabaster==0.7.10
anaconda-client==1.6.14
anaconda-navigator==1.8.7
anaconda-project==0.8.2
asn1crypto==0.24.0
astroid==1.6.3
astropy==3.0.2
astunparse==1.6.3
attrs==18.1.0
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.9.3
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
blis==0.4.1
bokeh==0.12.16
boto==2.48.0
boto3==1.17.26
botocore==1.20.26
Bottleneck==1.2.1
bs4==0.0.1
cachetools==4.2.1
catalogue==1.0.0
catboost==0.24.1
category-encoders==2.2.2
certifi==2020.12.5
cffi==1.11.5
chardet==3.0.4
cleantext==1.1.3
click==6.7
cloudpickle==1.3.0
clyent==1.2.2
colorama==0.4.4
colour==0.1.5
comtypes==1.1.4
conda==4.5.4
conda-build==3.10.5
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.2.2
cycler==0.10.0
cymem==2.0.5
Cython==0.29.14
cytoolz==0.9.0.1
dask==0.17.5
dataclasses==0.8
datashape==0.5.4
decorator==4.3.0
distributed==1.21.8
docutils==0.14
dtreeviz==1.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-2.3.1/en_core_web_lg-2.3.1.tar.gz
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
fastparquet==0.4.1
filelock==3.0.4
Flask==1.0.2
Flask-Cors==3.0.4
flatbuffers==1.12
gast==0.3.3
gensim==3.8.3
gevent==1.3.0
glob2==0.6
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
graphviz==0.17
greenlet==0.4.13
grpcio==1.32.0
h5py==2.10.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.6
imageio==2.3.0
imagesize==1.0.0
importlib-metadata==3.7.2
ipykernel==4.8.2
ipython==6.4.0
ipython_genutils==0.2.0
ipywidgets==7.2.1
isodate==0.6.0
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
jmespath==0.10.0
joblib==0.14.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lightgbm==3.0.0
llvmlite==0.23.1
locket==0.2.0
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==1.0
marshmallow==3.13.0
matplotlib==3.3.4
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
mljar==0.1.0
mljar-supervised==0.8.9
more-itertools==4.1.0
mpmath==1.0.0
msgpack-python==0.5.6
multipledispatch==0.5.0
murmurhash==1.0.5
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.19.5
numpydoc==0.8.0
oauthlib==3.1.0
odo==0.5.1
olefile==0.45.1
opencv-python==4.5.1.48
openpyxl==2.5.3
opt-einsum==3.3.0
packaging==17.1
pandas==1.1.2
pandocfilters==1.4.2
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.1
pep8==1.7.1
pickleshare==0.7.4
Pillow==8.3.1
pkginfo==1.4.2
plac==1.1.3
plotly==5.1.0
pluggy==0.6.0
ply==3.11
preshed==3.0.5
prompt-toolkit==1.0.15
protobuf==3.15.5
psutil==5.4.5
py==1.5.3
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.4
PyMySQL==1.0.2
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyreadstat==1.1.2
PySocks==1.7.1
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
pywin32==223
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
rdflib==5.0.0
regex==2020.11.13
requests==2.25.1
requests-futures==1.0.0
requests-oauthlib==1.3.0
rope==0.10.7
rsa==4.7.2
ruamel_yaml==0.15.35
s3transfer==0.3.4
sacremoses==0.0.43
scikit-image==0.13.1
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
sentencepiece==0.1.95
seqeval==1.2.2
shap==0.36.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
slicer==0.0.7
smart-open==4.2.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
soupsieve==2.2
spacy==2.3.2
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
srsly==1.0.5
statsmodels==0.12.2
stem==1.8.0
sympy==1.1.1
tables==3.4.3
tabulate==0.8.7
tblib==1.3.2
tenacity==8.0.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==7.4.1
threadpoolctl==2.1.0
thrift==0.13.0
tokenizers==0.5.2
toolz==0.9.0
torch==1.7.1+cpu
torchaudio==0.7.2
torchvision==0.8.2+cpu
tornado==5.0.2
torrequest==0.1.0
tqdm==4.59.0
traitlets==4.3.2
transformers==2.6.0
typing==3.6.4
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.26.3
wasabi==0.8.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.2.1
win-unicode-console==0.5
win_inet_pton==1.0.1
wincertstore==0.2
wordcloud==1.7.0
wrapt==1.12.1
xgboost==1.2.0
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.3.0
zict==0.1.3
zipp==3.4.1
